ai-service/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
uvicorn==0.24.0
aiohttp==3.9.1
